Add first working version of demo
This is the culmination of work at the AI4Life hackathon at
Human Technopole in Milan, Italy from 2023-02-06 thru 2023-02-10.

It includes a working Python-first demo of shared memory including
shared with an embedded Fiji/ImageJ application written in Java.

It also includes an unfinished rough draft of the Python
SharedMemory class translated into Java, so that Java code can make
use of the same mechanism (not backed by a file on disk) without
needing to go through a Python interpreter nor the imglyb
(https://github.com/imglib/imglyb) library like this demo does.

Finally, this repository contains some other unfinished notes,
experiments, and directions worked on during the hackathon, preserved
here so they are not forgotten, in case they are useful as a future
starting point for any of the elements described in the README.